An attempt to stabilize tests
