+++ ledger-core/v2/virtual/integration/stateunavailable_test.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/insolar/assured-ledger/ledger-core/v2/insolar/gen"
 	"github.com/insolar/assured-ledger/ledger-core/v2/insolar/payload"
-	"github.com/insolar/assured-ledger/ledger-core/v2/instrumentation/inslogger"
 	"github.com/insolar/assured-ledger/ledger-core/v2/reference"
 	"github.com/insolar/assured-ledger/ledger-core/v2/virtual/integration/utils"
 	"github.com/insolar/assured-ledger/ledger-core/v2/virtual/statemachine"
@@ -29,8 +28,9 @@ func makeVStateUnavailableEvent(t *testing.T, ref reference.Global, reason paylo
 
 func TestVirtual_VStateUnavailable_NoSuchObject(t *testing.T) {
 	t.Log("C4864")
-	server := utils.NewServerIgnoreLogErrors(t)
-	ctx := inslogger.TestContext(t)
+
+	server, ctx := utils.NewServerIgnoreLogErrors(t, nil)
+	defer server.Stop()
 
 	server.PublisherMock.Checker = func(topic string, messages ...*message.Message) error {
 		require.Len(t, messages, 1)
@@ -50,8 +50,9 @@ func TestVirtual_VStateUnavailable_NoSuchObject(t *testing.T) {
 
 func TestVirtual_VStateUnavailable_StateAlreadyExists(t *testing.T) {
 	t.Log("C4865")
-	server := utils.NewServerIgnoreLogErrors(t)
-	ctx := inslogger.TestContext(t)
+
+	server, ctx := utils.NewServerIgnoreLogErrors(t, nil)
+	defer server.Stop()
 
 	server.PublisherMock.Checker = func(topic string, messages ...*message.Message) error {
 		require.Len(t, messages, 1)
